move the tributaries to the right directory for comparison
